Backend_FarmerSocialNetwork/bin/requirements.txt
tensorflow==2.15.0
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
Pillow==8.3.2
